chromdiag/requirements.txt
colour-science
opencv-python
numpy